add Auditable::Base as abstract class for inheritence without breaking has_many association
